move the AppearanceStyle class to the package org.apache.pdfbox.pdmodel.interactive and use the same class in org.apache.pdfbox.pdmodel.interactive.form and org.apache.pdfbox.pdmodel.interactive.annotation.layout
